Port tests to be both xUnit and dnx raw program compatible
This let's us run the tests much faster and all from Visual Studio,
command line, and Appveyor.

TODO: Flag missing test methods
